# Minor refactoring