--- a/context.go
+++ b/context.go
@@ -4,18 +4,17 @@ import (
 	"context"
 )
 
-type key int
+// paramsKey represents the key for parameters
+type paramsKey struct{}
 
-const (
-	// ParamsKey is the key in a request context.
-	ParamsKey key = iota
-)
+// ParamsKey is the request context key under which URL params are stored.
+var ParamsKey = paramsKey{}
 
 // GetParam gets parameters from request.
 func GetParam(ctx context.Context, name string) string {
 	params, _ := ctx.Value(ParamsKey).([]Param)
 
-	for i := range params {
+	for i := 0; i < len(params); i++ {
 		if params[i].key == name {
 			return params[i].value
 		}