--- a/middleware.go
+++ b/middleware.go
@@ -10,14 +10,16 @@ type middleware func(http.Handler) http.Handler
 // middlewares represents the plural of middleware.
 type middlewares []middleware
 
+// NewMiddlewares creates a new middlewares.
 func NewMiddlewares(mws middlewares) middlewares {
 	return append([]middleware(nil), mws...)
 }
 
 // then executes middlewares.
 func (mws middlewares) then(h http.Handler) http.Handler {
-	for i := range mws {
-		h = mws[len(mws)-1-i](h)
+	l := len(mws)
+	for i := 0; i < l; i++ {
+		h = mws[l-1-i](h)
 	}
 	return h
 }