# Description
As a title.

# Changes
Some changes for improving performance.

before
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         9703232               119.8 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4622125               254.3 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2833615               426.2 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1898692               630.2 ns/op           352 B/op          5 allocs/op
BenchmarkWildCard5        498434              2323 ns/op             559 B/op          9 allocs/op
BenchmarkWildCard10       289339              4244 ns/op             801 B/op         14 allocs/op
BenchmarkRegexp1         1998573               600.1 ns/op           352 B/op          5 allocs/op
BenchmarkRegexp5          589863              2205 ns/op             559 B/op          9 allocs/op
BenchmarkRegexp10         310650              4092 ns/op             807 B/op         14 allocs/op
PASS
ok      github.com/bmf-san/goblin       13.604s
```

after
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1        10376299               119.2 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4794303               261.0 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2781861               428.4 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       2110044               567.3 ns/op           352 B/op          5 allocs/op
BenchmarkWildCard5        583000              2035 ns/op             559 B/op          9 allocs/op
BenchmarkWildCard10       338418              3643 ns/op             802 B/op         14 allocs/op
BenchmarkRegexp1         2148955               567.7 ns/op           352 B/op          5 allocs/op
BenchmarkRegexp5          602371              2026 ns/op             559 B/op          9 allocs/op
BenchmarkRegexp10         315440              3626 ns/op             801 B/op         14 allocs/op
PASS
ok      github.com/bmf-san/goblin       13.258s
```

# Impact range
No specification changes.

# Operational Requirements
N/A

# Related Issue
N/A

# Supplement
N/A